backend/requirements.txt
# Mosaico Backend - Requirements
# Modern stack 2025 - NO LangChain

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.1

# Google Cloud
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.0
google-api-python-client==2.149.0

# Authentication
clerk-backend-api==1.5.0

# Rate Limiting
slowapi==0.1.9

# Template Engine (for prompts)
jinja2==3.1.4

# Utils
python-dotenv==1.0.1
httpx==0.27.0

# Development
pytest==8.3.0
pytest-asyncio==0.24.0
